🛠️refactor: goalEditor 페이지 type 수정 - 상수로 관리하던 PRIVACY_TYPE 타입을 enum 타입으로 수정 - GoalEditorForm 컴포넌트 내부에서 정의된 타입을 상위 부모 goalEditor 디렉토리 내부 경로의 types.ts로 이동하여 관리되도록 수정 - 기존 GoalEditorFormInfo, GoalFormDataType 타입을 사용 중이던 파일들 경로에 맞게 수정
