+++ src/app/(route)/goal/editor/components/GoalEditor/components/GoalEditorForm/GoalEditorForm.constants.ts
@@ -1,5 +1,5 @@
-export const PRIVACY_TYPE = {
-  PUBLIC: "전체 공개",
-  FOLLOWER: "팔로워 공개",
-  PRIVATE: "나만보기",
-};
+export enum PRIVACY_TYPE {
+  PUBLIC = "전체 공개",
+  FOLLOWER = "팔로워 공개",
+  PRIVATE = "나만보기",
+}
