+++ src/app/(route)/goal/editor/goalEditor.types.ts
@@ -1,0 +1,12 @@
+import { GoalPrivacyType, RepeatDdudusType } from "@/app/_types/response/goal/goal";
+
+export interface GoalEditorFormInfo {
+  goal: string;
+}
+
+export interface GoalFormDataType {
+  goalText: string;
+  goalPrivacy: GoalPrivacyType;
+  color: string;
+  repeatDDuDu: RepeatDdudusType[];
+}
